+++ thebleep/specific/zypper.py
@@ -1,0 +1,3 @@
+from thebleep.utils import which
+
+zypper_available = bool(which('zypper'))